Grando_pro/requirements.txt
python-dotenv~=1.0.0
asyncpg~=0.27.0
aiogram~=2.25.1
asyncio~=3.4.3
#SQLAlchemy~=2.0.17
customtkinter~=5.2.0
Pillow~=10.0.0
